+++ agonyforge-mud-demo/src/main/java/com/agonyforge/mud/demo/cli/question/login/CharacterSpeciesQuestion.java
@@ -1,4 +1,4 @@
-package com.agonyforge.mud.demo.cli.question;
+package com.agonyforge.mud.demo.cli.question.login;
 
 import com.agonyforge.mud.core.cli.Color;
 import com.agonyforge.mud.core.cli.Question;
@@ -11,10 +11,12 @@ import com.agonyforge.mud.core.web.model.Input;
 import com.agonyforge.mud.core.web.model.Output;
 import com.agonyforge.mud.core.web.model.WebSocketContext;
 import com.agonyforge.mud.demo.cli.RepositoryBundle;
+
+import com.agonyforge.mud.demo.cli.question.BaseQuestion;
 import com.agonyforge.mud.demo.model.constant.Effort;
 import com.agonyforge.mud.demo.model.constant.Stat;
 import com.agonyforge.mud.demo.model.impl.MudCharacter;
-import com.agonyforge.mud.demo.model.impl.MudProfession;
+import com.agonyforge.mud.demo.model.impl.MudSpecies;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,19 +29,19 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
 
-import static com.agonyforge.mud.demo.model.impl.ModelConstants.TYPE_PROFESSION;
+import static com.agonyforge.mud.demo.model.impl.ModelConstants.TYPE_SPECIES;
 
 @Component
-public class CharacterProfessionQuestion extends BaseQuestion {
-    private static final Logger LOGGER = LoggerFactory.getLogger(CharacterProfessionQuestion.class);
+public class CharacterSpeciesQuestion extends BaseQuestion {
+    private static final Logger LOGGER = LoggerFactory.getLogger(CharacterSpeciesQuestion.class);
 
     private final MenuPane menuPane = new MenuPane();
 
     @Autowired
-    public CharacterProfessionQuestion(ApplicationContext applicationContext, RepositoryBundle repositoryBundle) {
+    public CharacterSpeciesQuestion(ApplicationContext applicationContext, RepositoryBundle repositoryBundle) {
         super(applicationContext, repositoryBundle);
 
-        menuPane.setTitle(new MenuTitle("Choose Your Profession"));
+        menuPane.setTitle(new MenuTitle("Choose Your Species"));
         menuPane.setPrompt(new MenuPrompt());
     }
 
@@ -54,7 +56,7 @@ public class CharacterProfessionQuestion extends BaseQuestion {
     public Response answer(WebSocketContext webSocketContext, Input input) {
         populateMenuItems();
 
-        String nextQuestion = "characterProfessionQuestion";
+        String nextQuestion = "characterSpeciesQuestion";
         Output output = new Output();
         String choice = input.getInput().toUpperCase();
         Optional<MenuItem> itemOptional = menuPane.getItems()
@@ -71,17 +73,17 @@ public class CharacterProfessionQuestion extends BaseQuestion {
 
             if (chOptional.isPresent()) {
                 MudCharacter ch = chOptional.get();
-                MudProfession profession = (MudProfession)item.getItem();
+                MudSpecies species = (MudSpecies)item.getItem();
 
-                ch.setProfessionId(profession.getId());
+                ch.setSpeciesId(species.getId());
 
-                Arrays.stream(Stat.values()).forEach(stat -> ch.setProfessionStat(stat, profession.getStat(stat)));
-                Arrays.stream(Effort.values()).forEach(effort -> ch.setProfessionEffort(effort, profession.getEffort(effort)));
+                Arrays.stream(Stat.values()).forEach(stat -> ch.setSpeciesStat(stat, species.getStat(stat)));
+                Arrays.stream(Effort.values()).forEach(effort -> ch.setSpeciesEffort(effort, species.getEffort(effort)));
 
                 getRepositoryBundle().getCharacterRepository().save(ch);
             }
 
-            nextQuestion = "characterMenuQuestion";
+            nextQuestion = "characterProfessionQuestion";
         }
 
         Question next = getQuestion(nextQuestion);
@@ -92,32 +94,32 @@ public class CharacterProfessionQuestion extends BaseQuestion {
     private void populateMenuItems() {
         menuPane.getItems().clear();
 
-        getRepositoryBundle().getProfessionRepository().getByType(TYPE_PROFESSION)
+        getRepositoryBundle().getSpeciesRepository().getByType(TYPE_SPECIES)
             .stream()
-            .sorted(Comparator.comparing(MudProfession::getName, String::compareToIgnoreCase))
-            .forEach(profession -> {
+            .sorted(Comparator.comparing(MudSpecies::getName, String::compareToIgnoreCase))
+            .forEach(species -> {
                 List<String> buffs = new ArrayList<>();
 
                 Arrays.stream(Stat.values())
-                    .filter(stat -> profession.getStat(stat) != 0)
-                    .forEach(stat -> buffs.add(String.format("%s %s%d",
-                        stat.getAbbreviation(),
-                        profession.getStat(stat) >= 0 ? "+" : "-",
-                        profession.getStat(stat))));
+                        .filter(stat -> species.getStat(stat) != 0)
+                        .forEach(stat -> buffs.add(String.format("%s %s%d",
+                            stat.getAbbreviation(),
+                            species.getStat(stat) >= 0 ? "+" : "-",
+                            species.getStat(stat))));
 
                 Arrays.stream(Effort.values())
-                    .filter(effort -> profession.getEffort(effort) != 0)
+                    .filter(effort -> species.getEffort(effort) != 0)
                     .forEach(effort -> buffs.add(String.format("%s %s%d",
                         effort.getName(),
-                        profession.getEffort(effort) >= 0 ? "+" : "-",
-                        profession.getEffort(effort))));
+                        species.getEffort(effort) >= 0 ? "+" : "-",
+                        species.getEffort(effort))));
 
                 menuPane.getItems().add(new MenuItem(
-                   Integer.toString(menuPane.getItems().size() + 1),
-                   // the formatting below is fragile and will need to be adjusted when you add
-                   // any profession with a name longer than "Human"
-                   String.format("%6s: %s", profession.getName(), String.join(", ", buffs)),
-                   profession
+                    Integer.toString(menuPane.getItems().size() + 1),
+                    // the formatting below is fragile and will need to be adjusted when you add
+                    // any species with a name longer than "Human"
+                    String.format("%6s: %s", species.getName(), String.join(", ", buffs)),
+                    species
                 ));
             });
     }
