+++ agonyforge-mud-demo/src/test/java/com/agonyforge/mud/demo/cli/question/login/CharacterMenuQuestionTest.java
@@ -1,4 +1,4 @@
-package com.agonyforge.mud.demo.cli.question;
+package com.agonyforge.mud.demo.cli.question.login;
 
 import com.agonyforge.mud.core.cli.Question;
 import com.agonyforge.mud.core.cli.Response;
